+++ src/Traits/IsPresentable.php
@@ -8,13 +8,12 @@ use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
 use ReflectionClass;
 use ReflectionMethod;
+use TPG\IsPresentable\Contracts\IsHidden;
 use TPG\IsPresentable\Exceptions\InvalidPresentableClass;
 use TPG\IsPresentable\Presenter;
 
 trait IsPresentable
 {
-    protected array $presentables = [];
-
     public function presentable(): Presenter
     {
         return new Presenter($this->getPresenters());
@@ -25,7 +24,7 @@ trait IsPresentable
         return array_merge(
             $this->getOriginalAttributes(),
             [
-                config('presentable.key') => $this->getPresenters(),
+                config('presentable.key') => $this->getPresenters(true),
             ],
         );
     }
@@ -39,13 +38,13 @@ trait IsPresentable
         return [];
     }
 
-    protected function getPresenters(): array
+    protected function getPresenters(bool $excludeHidden = false): array
     {
-        $presenters = collect($this->presenters)->merge($this->getPresentableMethods());
+        $presenters = collect($this->presenters)->merge($this->getPresenterMethods());
 
-        return $presenters->mapWithKeys(function (string|ReflectionMethod $value, $key) {
+        return $presenters->mapWithKeys(function (string|ReflectionMethod $value, $key) use ($excludeHidden) {
             if (is_string($value)) {
-                return [$key => $this->renderClass($value)];
+                return $this->renderClass($value, $key, $excludeHidden);
             }
 
             $name = Str::after($value->name, 'presentable');
@@ -54,7 +53,7 @@ trait IsPresentable
         })->toArray();
     }
 
-    protected function getPresentableMethods(): Collection
+    protected function getPresenterMethods(): Collection
     {
         $reflection = new ReflectionClass($this);
 
@@ -65,12 +64,16 @@ trait IsPresentable
             );
     }
 
-    protected function renderClass(string $className): string
+    protected function renderClass(string $className, string $attribute, bool $excludeIfHidden = false): array
     {
         if (! class_exists($className)) {
             throw new InvalidPresentableClass($className);
         }
 
-        return (new $className($this))?->render() ?? '';
+        if ($excludeIfHidden && (new ReflectionClass($className))->implementsInterface(IsHidden::class)) {
+            return [];
+        }
+
+        return [$attribute => (new $className($this))?->render() ?? ''];
     }
 }